Updated autonomouse to support gyros and updated tele-op
autonomous supports gyros now and tele-op has all the motors configured
and has some things commented out for testing.

+++ FtcRobotController/src/main/java/com/qualcomm/ftcrobotcontroller/opmodes/hydraDriveBase.java
@@ -7,6 +7,6 @@ public interface hydraDriveBase {
     void resetEncoders();
     void startMotors(double power1, double power2, double power3, double power4); //Assumes drivetrain has 4 motor drive
     void stopMotors();
-    double getEncoderAvg();
+    int getEncoderAvg();
 
 }
